gcp/a2a_agent/requirements.txt
click>=8.1.8
httpx>=0.28.1
langchain-google-genai
langchain-openai
langgraph>=1.0.0
pydantic>=2.10.6
python-dotenv>=1.1.0
uvicorn[standard]>=0.34.2
sse-starlette>=2.3.6
starlette>=0.46.2
a2a-sdk>=0.3.0
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter==1.0.0b42
langchain-azure-ai[opentelemetry]
